Add custom hook useJournalEntries as suggested in code review
